Create README.md
Measures of traits are the basis of functional biological diversity. Numerous works consider mean species‐level measures of traits while ignoring individual variance within species. However, there is a large amount of variation within species and it is increasingly apparent that it is important to consider trait variation not only between species, but also within species. Mammals are an interesting group for investigating trait‐based approaches because they play diverse and important ecological functions (e.g., pollination, seed dispersal, predation, grazing) that are correlated with functional traits. Here we compile a data set comprising morphological and life history information of 279 mammal species from 39,850 individuals of 388 populations ranging from −5.83 to −29.75 decimal degrees of latitude and −34.82 to −56.73 decimal degrees of longitude in the Atlantic forest of South America. We present trait information from 16,840 individuals of 181 species of non‐volant mammals (Rodentia, Didelphimorphia, Carnivora, Primates, Cingulata, Artiodactyla, Pilosa, Lagomorpha, Perissodactyla) and from 23,010 individuals of 98 species of volant mammals (Chiroptera). The traits reported include body mass, age, sex, reproductive stage, as well as the geographic coordinates of sampling for all taxa. Moreover, we gathered information on forearm length for bats and body length and tail length for rodents and marsupials. No copyright restrictions are associated with the use of this data set. Please cite this data paper when the data are used in publications. We also request that researchers and teachers inform us of how they are using the data.